requirement.txt
Django
requests
djangorestframework
httpretty
factory_boy
